backend/project/requirements/requirements.txt
Flask==1.0
gunicorn==19.9.0
Flask-SQLAlchemy
marshmallow